+++ R/manipulateWidget.R
@@ -29,14 +29,13 @@
 #'   corresponds to the Rstudio viewer, \code{"window"} to a dialog window, and
 #'   \code{"browser"} to an external web browser.
 #' @param .compare Sometimes one wants to compare the same chart but with two
-#'   different sets of parameters. This is the purpose of this argument. It must
-#'   be a named list whose names are the names of the inputs that should vary
-#'   between the two charts. All other parameters are common to the two charts
-#'   and changing their values will change the two charts. Each element of the
-#'   list must be a vector or a list of length 2 with the initial values of the
-#'   corresponding parameter for each chart. It can also be \code{NULL}. In this
-#'   case, the parameter is initialized with the default value for the two
-#'   charts.
+#'   different sets of parameters. This is the purpose of this argument. It can
+#'   be a character vector of input names or a named list whose names are the
+#'   names of the inputs that should vary between the two charts. Each element
+#'   of the list must be a vector or a list of length equal to the number of
+#'   charts with the initial values of the corresponding parameter for each
+#'   chart. It can also be \code{NULL}. In this case, the parameter is
+#'   initialized with the default value for the two charts.
 #' @param .compareOpts List of options created \code{\link{compareOptions}}.
 #'   These options indicate the number of charts to create and their disposition.
 #' @param .return A function that can be used to modify the output of
@@ -114,11 +113,21 @@
 #'     series2 = rnorm(100),
 #'     series3 = rnorm(100)
 #'   )
+#'
 #'   manipulateWidget(
 #'     dygraph(mydata[range[1]:range[2] - 2000, c("year", series)], main = title),
 #'     range = mwSlider(2001, 2100, c(2001, 2100)),
 #'     series = mwSelect(c("series1", "series2", "series3")),
 #'     title = mwText("Fictive time series"),
+#'     .compare = c("title", "series")
+#'   )
+#'
+#'   # Setting different initial values for each chart
+#'   manipulateWidget(
+#'     dygraph(mydata[range[1]:range[2] - 2000, c("year", series)], main = title),
+#'     range = mwSlider(2001, 2100, c(2001, 2100)),
+#'     series = mwSelect(c("series1", "series2", "series3")),
+#'     title = mwText(),
 #'     .compare = list(
 #'       title = list("First chart", "Second chart"),
 #'       series = NULL
@@ -216,16 +225,24 @@ manipulateWidget <- function(.expr, ..., .updateBtn = FALSE,
   .expr <- substitute(.expr)
   .viewer <- match.arg(.viewer)
   .env <- parent.frame()
-  compareMode <- !is.null(.compare)
+  compareMode <-
   .compareOpts <- do.call(compareOptions, .compareOpts)
 
-  if (compareMode) {
+  if (is.null(.compare)) {
+      .compareOpts$ncharts <- 1
+  } else {
+    if (is.character(.compare)) {
+      .compare <- match.arg(
+        .compare,
+        names(list(...)),
+        several.ok = TRUE
+      )
+      .compare <- sapply(.compare, function(x) NULL,
+                         simplify = FALSE, USE.NAMES = TRUE)
+    }
+
     if (is.null(.compareOpts$ncharts) || .compareOpts$ncharts < 2) {
       .compareOpts$ncharts <- 2
-    }
-  } else {
-    if (is.null(.compareOpts$ncharts)) {
-      .compareOpts$ncharts <- 1
     }
   }
 
@@ -257,7 +274,7 @@ manipulateWidget <- function(.expr, ..., .updateBtn = FALSE,
   dims <- .getRowAndCols(.compareOpts$ncharts, .compareOpts$nrow, .compareOpts$ncol)
 
   ui <- mwUI(controls, dims$nrow, dims$ncol, outputFunction, okBtn = !isRuntimeShiny,
-             updateBtn = .updateBtn)
+             updateBtn = .updateBtn, areaBtns = length(.compare) > 0)
   server <- mwServer(.expr, controls, initWidgets,
                      renderFunction,
                      .updateBtn,