+++ NEWS.md
@@ -8,6 +8,7 @@
 ## New features
 * `manipulateWidget()` now creates a more compact and elegant interface.
 * It is now possible to compare more than two charts. `manipulateWidget()` has a new argument `compareOpts` to control the number of charts and their position.
+* Argument `.compare` of `manipulateWidget` can now be a character vector. 
 
 # manipulateWidget 0.6.0 (2017-05-24)